W5D1/Q4/requirements.txt
flask==3.0.2
openai==1.12.0
python-dotenv==1.0.1
requests==2.31.0
markdown==3.5.2
python-slugify==8.0.4 